[OPENBUS-1185] Bugfix: no commit 128711 ao evitar o acúmulo de lixo ao baixar o arquivo de manifesto remoto, acabei criando um bug. Em repositório locais não se pode remover o manifesto! Um efeito colateral é que ao resolver as dependências o manifesto é essencial para sabermos o que já foi baixado e compilado previamente.
git-svn-id: https://subversion.tecgraf.puc-rio.br/engdist/openbus/puts/trunk@128717 ae0415b3-e90b-0410-900d-d0be9363c56b

+++ lua/tools/manifest.lua
@@ -123,7 +123,10 @@ function load(repo_url)
    end
 
    local manifest = local_loader(pathname)
-   os.remove(pathname)
+
+   if protocol ~= "file" then
+     os.remove(pathname)
+   end
    
    return manifest
 end
